functions/preprocessing/requirements.txt
mecab-python3
unidic-lite
pandas
numpy
faker